End Day

+++ Assets/Scripts/Abstractions/ValuesConsts/CellValues.cs
@@ -5,9 +5,9 @@
         public const byte CELL_COUNT_X = 15;
         public const byte CELL_COUNT_Y = 11;
 
-        public const int XY_FOR_ARRAY = 2;
-        public const int X = 0;
-        public const int Y = 1;
+        public const byte XY_FOR_ARRAY = 2;
+        public const byte X = 0;
+        public const byte Y = 1;
